requirements.txt

# Dependencias Oracle existentes
oracledb
python-dotenv
pandas
openpyxl
SQLAlchemy

# Dependencias SQL Server
pyodbc
pymssql

# Dependencias adicionales para logging y utilidades
argparse
logging